refactor: switch project to ESM
