+++ package.json
@@ -2,10 +2,11 @@
   "name": "notification-service",
   "version": "1.0.0",
   "main": "dist/index.js",
+  "type": "module",
   "scripts": {
     "init-hooks": "git config core.hooksPath .githooks",
-    "dev": "nodemon --exec \"ts-node --transpile-only --files\" src/index.ts",
-    "dev:workers": "nodemon --exec \"ts-node --transpile-only --files\" src/workers.ts",
+    "dev": "nodemon --watch src --ext ts --exec \"node --loader ts-node/esm\" src/index.ts",
+    "dev:workers": "nodemon --watch src --ext ts --exec \"node --loader ts-node/esm\" src/workers.ts",
     "build": "tsc",
     "start": "node dist/index.js",
     "start:workers": "node dist/workers.js",
