+++ src/__tests__/constants.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { ANONYMOUS_USER_ID, NOTIFICATION_STATUSES, isValidStatusId } from '../constants/';
+import { ANONYMOUS_USER_ID, NOTIFICATION_STATUSES, isValidStatusId } from '../constants/index.js';
 
 describe('constants', () => {
   describe('NOTIFICATION_STATUSES', () => {
