+++ src/config/logger.ts
@@ -1,5 +1,5 @@
 import winston from 'winston';
-import config from './env';
+import config from './env.js';
 
 const logger = winston.createLogger({
   level: config.logger.level,
