+++ src/exceptions/NotFoundError.ts
@@ -1,4 +1,4 @@
-import { BaseError } from './BaseError';
+import { BaseError } from './BaseError.js';
 
 /**
  * Resource not found
