+++ src/exceptions/ServiceUnavailableError.ts
@@ -1,4 +1,4 @@
-import { BaseError } from './BaseError';
+import { BaseError } from './BaseError.js';
 
 /**
  * External service unavailable (SMTP, RabbitMQ, Database)
