+++ src/exceptions/UnauthorizedError.ts
@@ -1,4 +1,4 @@
-import { BaseError } from './BaseError';
+import { BaseError } from './BaseError.js';
 
 /**
  * UnauthorizedError (missing or invalid api key)
