+++ src/exceptions/ValidationError.ts
@@ -1,4 +1,4 @@
-import { BaseError } from './BaseError';
+import { BaseError } from './BaseError.js';
 
 /**
  * Validation error (invalid data from client)
