+++ src/exceptions/index.ts
@@ -1,11 +1,11 @@
 /**
  * Centralized export for all custom exceptions
  */
-export { BaseError } from './BaseError';
-export { ValidationError } from './ValidationError';
-export { NotFoundError } from './NotFoundError';
-export { ServiceUnavailableError } from './ServiceUnavailableError';
-export { RateLimitError } from './RateLimitError';
-export { ConflictError } from './ConflictError';
-export { ForbiddenError } from './ForbiddenError';
-export { UnauthorizedError } from './UnauthorizedError';
+export { BaseError } from './BaseError.js';
+export { ValidationError } from './ValidationError.js';
+export { NotFoundError } from './NotFoundError.js';
+export { ServiceUnavailableError } from './ServiceUnavailableError.js';
+export { RateLimitError } from './RateLimitError.js';
+export { ConflictError } from './ConflictError.js';
+export { ForbiddenError } from './ForbiddenError.js';
+export { UnauthorizedError } from './UnauthorizedError.js';
