+++ src/helpers/index.ts
@@ -1,2 +1,2 @@
-export { default as httpClient } from './httpClient';
-export * from './validation';
+export { default as httpClient } from './httpClient.js';
+export * from './validation.js';
