+++ src/middleware/auth.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
-import config from '../config/env';
-import { UnauthorizedError } from '../exceptions/';
+import config from '../config/env.js';
+import { UnauthorizedError } from '../exceptions/index.js';
 
 export default function authenticate(req: Request, _res: Response, next: NextFunction) {
   if (!config.apiKey || req.path.startsWith('/api/health')) return next();
